Move function into a reusable util
